Review comments and moved config.go to context as it i think it makes a bit more sense.
