+++ context/config.go
@@ -11,12 +11,12 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package config
+package context
 
 import (
-  "github.com/sonatype-nexus-community/nancy/buildversion"
-	"os"
   "fmt"
+  "github.com/sonatype-nexus-community/nancy/buildversion"
+  "os"
   "path/filepath"
   "strings"
 )
