+++ go.mod
@@ -12,5 +12,5 @@ require (
 	github.com/sonatype-nexus-community/go-sona-types v0.0.8
 	github.com/sonatype-nexus-community/nancy v1.0.1
 	github.com/spf13/afero v1.3.4
-	gopkg.in/yaml.v2 v2.3.0
+	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776
 )
